Enable clang-tidy during CI process  (#1086)
* Enable clang-tidy during CI process and fix performance-* rule violations.

* Install clang-tidy package.

* Need to actually build to get all dependencies.

* Need to install minimum Python dependencies for linting.

* Build mimalloc too.

* Enable clang-analyzer rules.

* Enable concurrency rules.

* Add linting to macOS runner.

* Add macOS specific checks.

* Try different way of calling clang-tidy.

* Add clang-tidy to PATH.

* 'Build' using clang.

* Try LLVM 20.

* Use GitHub LLVM.

* LLVM 21 again.

* Use install-llvm action to install LLVM 20.

* Don't be so strict with version number.

* Suppress macOS specific lint errors.

* Fix missed errors.

* Fix Windows compiler error.

* Experiment: also enable clang-tidy for Windows build.

* Revert TcpConnectionHandler change.

* Another attempt at fixing Windows compiler error.

* Need to swap arguments on Windows.

* Add PR #1086 to changelog.

* Need run-clang-tidy on Windows.

* Disable clang-tidy for Windows builds.

* Remove commented code.

* Revert to master version of Windows CI script

* Try revert again

+++ src/pipeline/rade_text.c
@@ -301,7 +301,7 @@ void rade_text_rx(rade_text_t ptr, float *syms, int symSize)
             }
         }
     }
-    rms = sqrt(rms / symSize);
+    rms = sqrtf(rms / symSize);
 
     // Copy over symbols prior to decode.
     for (int index = 0; index < LDPC_TOTAL_SIZE_BITS / 2; index++)
